remove side effect from mi alias
It happens too often that I have to clean something up in the branch
soon after the first attempt to merge it in, and then I have to the
recreate the ref manually. I should just leave the ref around until
I know it won’t be needed any more.